// src/app/dashboard/models/production-plan.ts
export interface IProductionPlan {
    planPercent: number;
    resourceValue: number;
    stockValue: number;
    dataPercent: number;
    modelPercent: number;
}
export interface IProductionPlan {
    planPercent: number;
    factPercent: number;
    resourceValue: number;
    stockValue: number;
    dataPercent: number;
    modelPercent: number;
}